Move and split widget module
* Use backward compatible exports in package
